Difficulty: Easy & Max Score: 15
You will be given an integer n, and a threshold, k. For each number i from 1 through n, find the maximum value of the logical and, or and xor when compared against all integers through n that are greater than i. Consider a value only if the comparison returns a result less than k. Print the results of the and, or and exclusive or comparisons on separate lines, in that order.